+++ RELEASE-NOTES.md
@@ -38,6 +38,7 @@
 * Fixed an issue with Mono Initialization not rethrowing exceptions to logging in some rare cases
 * Implemented experimental MonoBleedingEdge Environment Patches behind Loader config option (Default is OFF)
 * Fixed an issue with Harmony Auto-Patching for Plugins and Mods not gracefully failing
+* Implemented Harmony TryPatchAll Extension to MelonUtils
 
 ## New Contributors
 * [Squaduck](<https://github.com/Squaduck>) made their first contribution in [#970](<https://github.com/LavaGang/MelonLoader/pull/970>)